ui: split flamegraph into tree explorer building blocks
The flamegraph widget currently owns everything: the filter bar, the
measure picker, highlight state, export, the canvas and the fetch
orchestration wrapped around it in QueryFlamegraph. Upcoming work adds
top-down/bottom-up call tree and flat function table views that share
the same filters and fetched data, which is not possible while all of
this is fused together.

Decompose it into:

- widgets/tree_explorer.ts: the shared state schema, filter grammar,
  data types and formatting helpers, renamed to TreeExplorer* since
  they now describe the whole family of views. Serialized field names
  are unchanged so persisted state keeps parsing. The old
  Flamegraph.createDefaultState / updateState statics become free
  functions here.
- widgets/tree_explorer_filter_bar.ts: the filter / measure /
  highlight bar as a standalone widget. Highlight is lifted into attrs
  so sibling views can consume the same pattern; export content is
  provided by the caller so it can later track the active view.
- widgets/flamegraph.ts: now only the canvas and its tooltip.
- components/tree_explorer_fetcher.ts: the mithril-free data layer
  that was QueryFlamegraph (virtual table cache + tree query).
- components/tree_explorer_panel.ts: the batteries-included
  composition replacing FlamegraphPanel, owning the fetcher lifecycle
  and rendering bar plus flamegraph.

This is a no-op refactor apart from one small cleanup: the redundant
"Hide" button in the highlight row is dropped, since the Highlight
toggle already closes it.
