+++ ui/src/plugins/com.android.DayExplorer/index.ts
@@ -22,12 +22,15 @@ import {TrackNode} from '../../public/workspace';
 import {STR, LONG, LONG_NULL} from '../../trace_processor/query_result';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {type AreaSelection, areaSelectionsEqual} from '../../public/selection';
-import {Flamegraph, FLAMEGRAPH_STATE_SCHEMA} from '../../widgets/flamegraph';
+import {
+  TREE_EXPLORER_STATE_SCHEMA,
+  updateTreeExplorerState,
+} from '../../widgets/tree_explorer';
 import {
   metricsFromTableOrSubquery,
-  type QueryFlamegraphMetric,
-} from '../../components/query_flamegraph';
-import {FlamegraphPanel} from '../../components/flamegraph_panel';
+  type TreeExplorerQueryMetric,
+} from '../../components/tree_explorer_fetcher';
+import {TreeExplorerPanel} from '../../components/tree_explorer_panel';
 import SupportPlugin from '../com.android.AndroidLongBatterySupport';
 import type {Store} from '../../base/store';
 import {z} from 'zod';
@@ -36,7 +39,7 @@ import {ensureExists} from '../../base/assert';
 const DAY_EXPLORER_TRACK_KIND = 'day_explorer_counter_track';
 
 const DAY_EXPLORER_PLUGIN_STATE_SCHEMA = z.object({
-  areaSelectionFlamegraphState: FLAMEGRAPH_STATE_SCHEMA.optional(),
+  areaSelectionFlamegraphState: TREE_EXPLORER_STATE_SCHEMA.optional(),
 });
 
 type DayExplorerPluginState = z.infer<typeof DAY_EXPLORER_PLUGIN_STATE_SCHEMA>;
@@ -145,7 +148,7 @@ export default class DayExplorerPlugin implements PerfettoPlugin {
 
   private createDayExplorerFlameGraphPanel(trace: Trace) {
     let previousSelection: AreaSelection | undefined;
-    let flamegraphMetrics: ReadonlyArray<QueryFlamegraphMetric> | undefined;
+    let flamegraphMetrics: ReadonlyArray<TreeExplorerQueryMetric> | undefined;
     return {
       id: 'day_explorer_flamegraph_selection',
       name: 'Day Explorer Flamegraph',
@@ -163,7 +166,7 @@ export default class DayExplorerPlugin implements PerfettoPlugin {
         const store = ensureExists(this.store);
         return {
           isLoading: false,
-          content: m(FlamegraphPanel, {
+          content: m(TreeExplorerPanel, {
             trace,
             metrics: flamegraphMetrics,
             state: store.state.areaSelectionFlamegraphState,
@@ -180,7 +183,7 @@ export default class DayExplorerPlugin implements PerfettoPlugin {
 
   private computeDayExplorerFlameGraph(
     currentSelection: AreaSelection,
-  ): ReadonlyArray<QueryFlamegraphMetric> | undefined {
+  ): ReadonlyArray<TreeExplorerQueryMetric> | undefined {
     // The flame graph will be shown when any day explorer track is in the area
     // selection. The selection is used to filter by time, and we filter the graph
     // to only include energy from the selected tracks and their recursive descendants.
@@ -268,7 +271,7 @@ export default class DayExplorerPlugin implements PerfettoPlugin {
     });
     const store = ensureExists(this.store);
     store.edit((draft) => {
-      draft.areaSelectionFlamegraphState = Flamegraph.updateState(
+      draft.areaSelectionFlamegraphState = updateTreeExplorerState(
         draft.areaSelectionFlamegraphState,
         metrics,
       );
