+++ ui/src/plugins/dev.perfetto.HeapProfile/heap_profile_track.ts
@@ -20,7 +20,7 @@ import {SliceTrack} from '../../components/tracks/slice_track';
 import type {Trace} from '../../public/trace';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {LONG, LONG_NULL, NUM, STR} from '../../trace_processor/query_result';
-import type {FlamegraphState} from '../../widgets/flamegraph';
+import type {TreeExplorerState} from '../../widgets/tree_explorer';
 import {profileDescriptor} from './common';
 import {HeapProfileFlamegraphDetailsPanel} from './heap_profile_details_panel';
 
@@ -30,8 +30,8 @@ export function createHeapProfileTrack(
   tableName: string,
   upid: number,
   heapProfileIsIncomplete: boolean,
-  detailsPanelState: FlamegraphState | undefined,
-  onDetailsPanelStateChange: (state: FlamegraphState) => void,
+  detailsPanelState: TreeExplorerState | undefined,
+  onDetailsPanelStateChange: (state: TreeExplorerState) => void,
   onNodeSelected?: (args: {
     pathHashes: string;
     isDominator: boolean;
