+++ ui/src/plugins/dev.perfetto.TraceProcessorTrack/index.ts
@@ -27,10 +27,10 @@ import {sliceDistributionCellRenderers} from '../../components/details/slice_det
 import {openDistributionTab} from '../../components/distribution_panel';
 import {
   metricsFromTableOrSubquery,
-  type QueryFlamegraphDependency,
-  type QueryFlamegraphMetric,
-} from '../../components/query_flamegraph';
-import {FlamegraphPanel} from '../../components/flamegraph_panel';
+  type TreeExplorerFetcherDependency,
+  type TreeExplorerQueryMetric,
+} from '../../components/tree_explorer_fetcher';
+import {TreeExplorerPanel} from '../../components/tree_explorer_panel';
 import type {MinimapRow} from '../../public/minimap';
 import type {PerfettoPlugin} from '../../public/plugin';
 import {type AreaSelection, areaSelectionsEqual} from '../../public/selection';
@@ -48,7 +48,10 @@ import {
   STR_NULL,
 } from '../../trace_processor/query_result';
 import {escapeSearchQuery} from '../../trace_processor/query_utils';
-import {Flamegraph, FLAMEGRAPH_STATE_SCHEMA} from '../../widgets/flamegraph';
+import {
+  TREE_EXPLORER_STATE_SCHEMA,
+  updateTreeExplorerState,
+} from '../../widgets/tree_explorer';
 import ProcessThreadGroupsPlugin from '../dev.perfetto.ProcessThreadGroups';
 import StandardGroupsPlugin from '../dev.perfetto.StandardGroups';
 import {CounterSelectionAggregator} from './counter_selection_aggregator';
@@ -70,7 +73,7 @@ import {ThreadSliceDetailsPanel} from '../../components/details/thread_slice_det
 import {CallstackDetailsSection} from './callstack_details_section';
 
 const TRACE_PROCESSOR_TRACK_PLUGIN_STATE_SCHEMA = z.object({
-  areaSelectionFlamegraphState: FLAMEGRAPH_STATE_SCHEMA.optional(),
+  areaSelectionFlamegraphState: TREE_EXPLORER_STATE_SCHEMA.optional(),
 });
 
 type TraceProcessorTrackPluginState = z.infer<
@@ -78,8 +81,8 @@ type TraceProcessorTrackPluginState = z.infer<
 >;
 
 interface SliceFlamegraphData extends AsyncDisposable {
-  readonly metrics: ReadonlyArray<QueryFlamegraphMetric>;
-  readonly dependencies: ReadonlyArray<QueryFlamegraphDependency>;
+  readonly metrics: ReadonlyArray<TreeExplorerQueryMetric>;
+  readonly dependencies: ReadonlyArray<TreeExplorerFetcherDependency>;
 }
 
 function createDetailsPanel(trace: Trace, utid: number | null) {
@@ -643,7 +646,7 @@ export default class TraceProcessorTrackPlugin implements PerfettoPlugin {
           isLoading,
           content:
             computed &&
-            m(FlamegraphPanel, {
+            m(TreeExplorerPanel, {
               trace,
               metrics: computed.metrics,
               dependencies: computed.dependencies,
@@ -784,14 +787,13 @@ export default class TraceProcessorTrackPlugin implements PerfettoPlugin {
     });
     const store = ensureExists(this.store);
     store.edit((draft) => {
-      draft.areaSelectionFlamegraphState = Flamegraph.updateState(
+      draft.areaSelectionFlamegraphState = updateTreeExplorerState(
         draft.areaSelectionFlamegraphState,
         metrics,
       );
     });
-    const dependency: QueryFlamegraphDependency = SharedAsyncDisposable.wrap(
-      disposables.move(),
-    );
+    const dependency: TreeExplorerFetcherDependency =
+      SharedAsyncDisposable.wrap(disposables.move());
     return {
       metrics,
       dependencies: [dependency],
