+++ ui/src/widgets/flamegraph.scss
@@ -24,84 +24,6 @@
     height: 100%;
   }
 
-  .filter-bar {
-    margin: 6px 8px;
-    display: flex;
-    column-gap: 6px;
-    row-gap: 6px;
-    align-items: center;
-    flex-wrap: wrap;
-
-    > .pf-tag-input {
-      flex-grow: 1;
-    }
-
-    .pf-select {
-      flex-shrink: 0;
-    }
-
-    .pf-radio-group__button {
-      padding-block: 1px;
-    }
-  }
-
-  .pf-tag-input {
-    flex: 1 1 0;
-    min-width: 0;
-    max-height: 60px;
-    overflow-y: auto;
-    flex-wrap: wrap;
-    gap: 6px;
-
-    // Override the inner Stack to allow wrapping and add gaps.
-    .pf-stack {
-      flex-wrap: wrap;
-      gap: 6px;
-    }
-
-    .pf-chip {
-      max-width: 200px;
-      min-width: auto;
-    }
-
-    input {
-      min-width: 80px;
-    }
-  }
-
-  .pf-flamegraph-secondary-row {
-    flex: 1 0 100%;
-    min-width: 0;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 6px;
-  }
-
-  .pf-flamegraph-control-label {
-    font-weight: 500;
-    flex-shrink: 0;
-  }
-
-  .pf-flamegraph-highlight-search {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    flex: 0 1 416px;
-    min-width: 280px;
-
-    .pf-text-input {
-      min-width: 180px;
-      width: 100%;
-    }
-  }
-
-  .pf-flamegraph-highlight-search__count {
-    flex-shrink: 0;
-    color: var(--pf-color-text-muted);
-    font-variant-numeric: tabular-nums;
-  }
-
   .popup-anchor {
     width: 0px;
     height: 0px;
@@ -155,60 +77,6 @@
     font-style: italic;
     color: var(--pf-color-text-muted);
     padding-bottom: 4px; // Extra gap between marker and function name
-  }
-}
-
-.pf-filter-builder {
-  &.pf-popup {
-    min-width: 300px;
-    > .pf-popup-content {
-      padding: 12px;
-    }
-  }
-  &__separator {
-    height: 1px;
-    background: var(--pf-color-border);
-    margin: 12px 0 8px 0;
-  }
-  &__tip {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    font-size: smaller;
-    color: var(--pf-color-text-muted);
-    padding-bottom: 8px;
-  }
-  &__help {
-    font-size: 13px;
-  }
-  &__help-title {
-    font-weight: 500;
-    margin-bottom: 4px;
-  }
-  &__help-row {
-    margin: 4px 0;
-  }
-  &__desc,
-  &__warn,
-  &__hint {
-    font-size: smaller;
-    margin-bottom: 8px;
-  }
-  &__desc,
-  &__hint {
-    color: var(--pf-color-text-muted);
-  }
-  &__hint {
-    margin-top: 4px;
-  }
-  code {
-    font-family: var(--pf-font-monospace, monospace);
-    background: var(--pf-color-background-secondary, rgba(0, 0, 0, 0.06));
-    border-radius: 3px;
-    padding: 0 3px;
-  }
-  &__warn {
-    color: var(--pf-color-warning);
   }
 }
