+++ ui/src/widgets/flamegraph.ts
@@ -15,42 +15,46 @@
 import './flamegraph.scss';
 import m from 'mithril';
 import {ensureExists, assertTrue, assertUnreachable} from '../base/assert';
-import {fuzzySearch} from '../base/fuzzy';
 import {Monitor} from '../base/monitor';
 import {Button, ButtonBar} from './button';
-import {Chip} from './chip';
-import {Intent} from './common';
 import {copyToClipboard} from '../base/clipboard';
-import {CopyToClipboardButton} from './copy_to_clipboard_button';
 import {EmptyState} from './empty_state';
-import {ExportButton, type ExportFormat} from './export_button';
+import type {ExportFormat} from './export_button';
 import {
   formatAsTSV,
   formatAsJSON,
   formatAsMarkdown,
 } from '../base/export_formatters';
-import {Form, FormLabel} from './form';
-import {Icon} from './icon';
-import {MiddleEllipsis} from './middle_ellipsis';
 import {Popup, PopupPosition} from './popup';
-import {RadioGroup} from './radio_group';
-import {Select} from './select';
 import {Spinner} from './spinner';
-import {TagInput} from './tag_input';
-import {TextInput} from './text_input';
-import {Tooltip} from './tooltip';
-import {z} from 'zod';
 import type {Rect2D, Size2D} from '../base/geom';
 import {
   VirtualOverlayCanvas,
   type VirtualOverlayCanvasApi,
 } from './virtual_overlay_canvas';
-import {MenuDivider, MenuItem, type MenuItemAttrs, PopupMenu} from './menu';
+import {MenuItem, type MenuItemAttrs, PopupMenu} from './menu';
 import {type Color, HSLColor} from '../base/color';
 import {hash} from '../base/hash';
-import {escapeRegex, parseUserFilterRegex} from './flamegraph_regex';
+import {escapeRegex} from './flamegraph_regex';
 import type {MithrilEvent} from '../base/mithril_utils';
 import {Icons} from '../base/semantic_icons';
+import {
+  type ActionCategory,
+  FILTER_TYPES,
+  type FilterType,
+  type TreeExplorerData,
+  type TreeExplorerFilter,
+  type TreeExplorerMetric,
+  type TreeExplorerNode,
+  type TreeExplorerOptionalAction,
+  type TreeExplorerPropertyDefinition,
+  type TreeExplorerState,
+  addFilter,
+  displayPercentage,
+  displaySize,
+  getUnitDisplayName,
+  metricId,
+} from './tree_explorer';
 
 const LABEL_FONT_STYLE = '12px Roboto';
 const NODE_HEIGHT = 20;
@@ -94,32 +98,6 @@ interface ZoomRegion {
   readonly type: 'ABOVE_ROOT' | 'BELOW_ROOT' | 'ROOT';
 }
 
-// Context passed to a FlamegraphOptionalAction's execute callback.
-//
-// `properties` is the (reduced) kv map of the user-declared
-// unaggregatableProperties / aggregatableProperties on the metric.
-//
-// `node` is the clicked flamegraph node for node-level actions, and undefined
-// for root-level actions (where there is no specific node).
-export interface FlamegraphActionContext {
-  readonly properties: ReadonlyMap<string, string>;
-  readonly node?: FlamegraphNode;
-}
-
-export interface FlamegraphOptionalAction {
-  readonly name: string;
-  execute?: (ctx: FlamegraphActionContext) => void;
-  readonly subActions?: FlamegraphOptionalAction[];
-  // Presentation in the categorized node menu; absent category → "Drill down".
-  readonly icon?: string;
-  readonly description?: m.Children;
-  readonly category?: ActionCategory;
-}
-
-// FOCUS re-frames without removing data; FILTER reshapes what's shown; DRILL
-// inspects elsewhere; COPY exports.
-export type ActionCategory = 'FOCUS' | 'FILTER' | 'DRILL' | 'COPY';
-
 const CATEGORY_LABELS: Record<ActionCategory, string> = {
   FOCUS: 'Focus',
   FILTER: 'Filter',
@@ -149,342 +127,30 @@ interface NodeAction {
   execute(): void;
 }
 
-export interface FlamegraphOptionalMarker {
-  readonly name: string;
-  isVisible: (properties: ReadonlyMap<string, string>) => boolean;
-}
-
-export type FlamegraphPropertyDefinition = {
-  displayName: string;
-  value: string;
-  isVisible: boolean;
-  isAggregatable: boolean;
-};
-
-export interface FlamegraphNode {
-  readonly id: number;
-  readonly parentId: number;
-  readonly depth: number;
-  readonly name: string;
-  readonly selfValue: number;
-  readonly cumulativeValue: number;
-  readonly parentCumulativeValue?: number;
-  readonly properties: ReadonlyMap<string, FlamegraphPropertyDefinition>;
-  readonly marker?: string;
-  readonly xStart: number;
-  readonly xEnd: number;
-}
-
-export interface FlamegraphQueryData {
-  readonly nodes: ReadonlyArray<FlamegraphNode>;
-  readonly unfilteredCumulativeValue: number;
-  readonly allRootsCumulativeValue: number;
-  readonly minDepth: number;
-  readonly maxDepth: number;
-  readonly nodeActions: ReadonlyArray<FlamegraphOptionalAction>;
-  readonly rootActions: ReadonlyArray<FlamegraphOptionalAction>;
-}
-
-const FLAMEGRAPH_FILTER_SCHEMA = z
-  .object({
-    kind: z
-      .union([
-        z.literal('SHOW_STACK').readonly(),
-        z.literal('HIDE_STACK').readonly(),
-        z.literal('HIDE_FRAME').readonly(),
-        z.literal('OPTIONS').readonly(),
-      ])
-      .readonly(),
-    filter: z.string().readonly(),
-  })
-  .readonly();
-
-type FlamegraphFilter = z.infer<typeof FLAMEGRAPH_FILTER_SCHEMA>;
-
-const FLAMEGRAPH_VIEW_SCHEMA = z
-  .discriminatedUnion('kind', [
-    z.object({kind: z.literal('TOP_DOWN').readonly()}),
-    z.object({kind: z.literal('BOTTOM_UP').readonly()}),
-    z.object({
-      kind: z.literal('FROM_FRAME').readonly(),
-      pattern: z.string().readonly(),
-      displayLabel: z.string().optional().readonly(),
-    }),
-    z.object({
-      kind: z.literal('PIVOT').readonly(),
-      pivot: z.string().readonly(),
-      // Display text for the pivot chip; SQL match still uses `pivot`.
-      displayLabel: z.string().optional().readonly(),
-    }),
-  ])
-  .readonly();
-
-export type FlamegraphView = z.infer<typeof FLAMEGRAPH_VIEW_SCHEMA>;
-
-export const FLAMEGRAPH_STATE_SCHEMA = z
-  .object({
-    selectedMetricId: z.string().readonly(),
-    addedMetricIds: z.array(z.string()).default([]),
-    filters: z.array(FLAMEGRAPH_FILTER_SCHEMA),
-    view: FLAMEGRAPH_VIEW_SCHEMA,
-  })
-  .readonly();
-
-export type FlamegraphState = z.infer<typeof FLAMEGRAPH_STATE_SCHEMA>;
-
-interface FlamegraphMetric {
-  // Stable identity used in persisted state. Defaults to `name`.
-  readonly id?: string;
-  readonly name: string;
-  readonly unit: string;
-  // Where the measure came from. Undefined is treated as ADDED so callers
-  // only need to mark the small set of built-in defaults.
-  readonly provenance?: 'DEFAULT' | 'ADDED';
-  // Label for the name column in copy stack table and tooltip.
-  // Examples: "Symbol", "Slice", "Class". Defaults to "Name".
-  readonly nameColumnLabel?: string;
-}
-
-export interface FlamegraphAddableMetric {
-  readonly id: string;
-  readonly name: string;
-}
-
 export interface FlamegraphAttrs {
-  readonly metrics: ReadonlyArray<FlamegraphMetric>;
-  readonly state: FlamegraphState;
-  readonly data: FlamegraphQueryData | undefined;
-  readonly addableMetrics?: ReadonlyArray<FlamegraphAddableMetric>;
+  readonly metrics: ReadonlyArray<TreeExplorerMetric>;
+  readonly state: TreeExplorerState;
+  readonly data: TreeExplorerData | undefined;
 
-  readonly onStateChange: (filters: FlamegraphState) => void;
-  readonly onAddMetric?: (metric: FlamegraphAddableMetric) => void;
-}
+  // Nodes whose name matches are drawn with a highlighted border. Owned by
+  // the caller so the (separate) filter bar can drive it.
+  readonly highlightRegex?: RegExp;
 
-type FilterType =
-  'SHOW_STACK' | 'HIDE_STACK' | 'SHOW_FROM_FRAME' | 'HIDE_FRAME' | 'PIVOT';
-type PatternViewKind = 'FROM_FRAME' | 'PIVOT';
-
-interface FilterTypeOption {
-  readonly value: FilterType;
-  // Canonical name; also a valid filter-bar syntax prefix.
-  readonly label: string;
-  readonly friendlyLabel: string;
-  readonly shortLabel: string;
-  readonly icon: string;
-  readonly category: ActionCategory;
-  readonly description: string;
-  // Example pattern used in tips for this filter type.
-  readonly example: string;
-  // Name used by other profilers, if any; surfaced in the node menu.
-  readonly aka?: string;
-}
-
-const FILTER_TYPES: ReadonlyArray<FilterTypeOption> = [
-  {
-    value: 'SHOW_STACK',
-    label: 'Show Stack',
-    friendlyLabel: 'Keep stacks matching name',
-    shortLabel: 'SS',
-    example: 'HandleRequest',
-    icon: 'visibility',
-    category: 'FILTER',
-    description:
-      'Keep only samples whose stack contains a frame whose name matches.',
-  },
-  {
-    value: 'HIDE_STACK',
-    label: 'Hide Stack',
-    friendlyLabel: 'Hide stacks matching name',
-    shortLabel: 'HS',
-    example: 'malloc',
-    icon: 'visibility_off',
-    category: 'FILTER',
-    description:
-      'Remove samples whose stack contains a frame whose name matches.',
-    aka: 'Drop function',
-  },
-  {
-    value: 'SHOW_FROM_FRAME',
-    label: 'Show From Frame',
-    friendlyLabel: 'Show from matching frame',
-    shortLabel: 'SFF',
-    example: 'HandleRequest',
-    icon: 'center_focus_strong',
-    category: 'FOCUS',
-    description:
-      'Re-root at matching frames and show their descendants, dropping ancestors.',
-    aka: 'Focus on subtree',
-  },
-  {
-    value: 'HIDE_FRAME',
-    label: 'Hide Frame',
-    friendlyLabel: 'Merge matching frames into caller',
-    shortLabel: 'HF',
-    example: '/.*alloc.*/i',
-    icon: 'call_merge',
-    category: 'FILTER',
-    description:
-      'Remove frames whose name matches, merging their children into the caller.',
-    aka: 'Merge function',
-  },
-  {
-    value: 'PIVOT',
-    label: 'Pivot',
-    friendlyLabel: 'Pivot on matching frames',
-    shortLabel: 'P',
-    example: '/.*alloc.*/i',
-    icon: 'account_tree',
-    category: 'FOCUS',
-    description:
-      'Re-root at matching frames with callers above and callees below.',
-  },
-];
-
-interface FilterBuilderAttrs {
-  readonly activePatternView?: PatternViewKind;
-  readonly onAdd: (filters: Array<{type: FilterType; value: string}>) => void;
-}
-
-class FilterBuilder implements m.ClassComponent<FilterBuilderAttrs> {
-  private type: FilterType = 'SHOW_STACK';
-  private filter = '';
-
-  view({attrs}: m.CVnode<FilterBuilderAttrs>) {
-    const {onAdd} = attrs;
-    const opt = FILTER_TYPES.find((o) => o.value === this.type);
-    const replacesPatternView =
-      (this.type === 'SHOW_FROM_FRAME' || this.type === 'PIVOT') &&
-      attrs.activePatternView !== undefined;
-
-    return m(
-      Form,
-      {
-        submitLabel: 'Add',
-        cancelLabel: 'Cancel',
-        onSubmit: () => {
-          if (!this.filter.trim()) return;
-          onAdd([{type: this.type, value: this.filter.trim()}]);
-          this.filter = '';
-        },
-        validation: () => this.filter.trim() !== '',
-      },
-      m(FormLabel, 'Type'),
-      m(
-        Select,
-        {
-          oninput: (e: Event) => {
-            this.type = (e.target as HTMLSelectElement).value as FilterType;
-          },
-        },
-        FILTER_TYPES.map((o) => m('option', {value: o.value}, o.friendlyLabel)),
-      ),
-      opt && m('.pf-filter-builder__desc', opt.description),
-      m(FormLabel, 'Filter'),
-      m(TextInput, {
-        autofocus: true,
-        placeholder: 'e.g. malloc',
-        value: this.filter,
-        onInput: (value) => {
-          this.filter = value;
-        },
-      }),
-      m(
-        '.pf-filter-builder__hint',
-        'Bare text matches literally and case-insensitively (e.g. ',
-        m('code', 'malloc'),
-        '). Use ',
-        m('code', '/…/'),
-        ' for a case-sensitive regex (e.g. ',
-        m('code', '/.*Alloc.*/'),
-        '), or append ',
-        m('code', 'i'),
-        ' for a case-insensitive regex (e.g. ',
-        m('code', '/.*alloc.*/i'),
-        ').',
-      ),
-      replacesPatternView &&
-        m(
-          '.pf-filter-builder__warn',
-          `Replaces the current ${
-            attrs.activePatternView === 'PIVOT' ? 'Pivot' : 'Show From Frame'
-          } filter.`,
-        ),
-      m('.pf-filter-builder__separator'),
-      opt &&
-        m(
-          '.pf-filter-builder__tip',
-          m(Icon, {icon: 'lightbulb_outline'}),
-          ' Tip: type ',
-          m('code', `${opt.shortLabel}: ${opt.example}`),
-          ' directly in the filter bar ',
-          m(
-            Tooltip,
-            {trigger: m(Icon, {icon: 'help_outline'})},
-            m(
-              '.pf-filter-builder__help',
-              m(
-                '.pf-filter-builder__help-title',
-                'Filter bar syntax (bare text is case-insensitive; /…/ is a ' +
-                  'case-sensitive regex and /…/i is case-insensitive):',
-              ),
-              FILTER_TYPES.map((o) =>
-                m(
-                  '.pf-filter-builder__help-row',
-                  m('strong', `${o.shortLabel}:`),
-                  ` ${o.label}, e.g. `,
-                  m('code', `${o.shortLabel}: ${o.example}`),
-                ),
-              ),
-              m(
-                '.pf-filter-builder__help-row',
-                'Combine operations by separating them with spaces, e.g. ',
-                m('code', 'SS: HandleRequest HF: /.*alloc.*/i'),
-              ),
-            ),
-          ),
-        ),
-    );
-  }
+  readonly onStateChange: (state: TreeExplorerState) => void;
 }
 
 /*
- * Widget for visualizing "tree-like" data structures using an interactive
- * flamegraph visualization.
- *
- * To use this widget, provide an array of "metrics", which correspond to
- * different properties of the tree to switch between (e.g. object size
- * and object count) and the data which should be displayed.
+ * Widget visualizing "tree-like" data structures using an interactive
+ * flamegraph visualization: just the canvas and its tooltip. The filtering
+ * bar lives in the separate TreeExplorerFilterBar widget; most callers should
+ * use TreeExplorerPanel (in components/) which composes the two and fetches
+ * the data.
  *
  * Note that it's valid to pass "undefined" as the data: this will cause a
  * loading container to be shown.
- *
- * Example:
- *
- * ```
- * const metrics = [...];
- * let state = ...;
- * let data = ...;
- *
- * m(Flamegraph, {
- *   metrics,
- *   state,
- *   data,
- *   onStateChange: (newState) => {
- *     state = newState,
- *     data = undefined;
- *     fetchData();
- *   },
- * });
- * ```
  */
 export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   private attrs: FlamegraphAttrs;
-
-  private showFilterBuilder: boolean = false;
-  private quickAddValue: string = '';
-  private showHighlightSearch = false;
-  private highlightPattern = '';
-  private highlightRegex?: RegExp;
 
   private dataChangeMonitor = new Monitor([() => this.attrs.data]);
   private zoomRegion?: ZoomRegion;
@@ -533,7 +199,12 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
     if (this.pendingScrollToY === undefined || this.canvasApi === undefined) {
       return;
     }
-    this.canvasApi.scrollTo({y: this.pendingScrollToY});
+    const requestedY = this.pendingScrollToY;
+    const {y: appliedY} = this.canvasApi.scrollTo({y: requestedY});
+    if (requestedY !== Number.MAX_SAFE_INTEGER && appliedY !== requestedY) {
+      // Keep shallow graphs aligned to whole rows when the target is unreachable.
+      this.canvasApi.scrollTo({y: 0});
+    }
     this.pendingScrollToY = undefined;
   }
 
@@ -561,7 +232,6 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
       this.canvasApi = undefined;
       return m(
         '.pf-flamegraph',
-        this.renderFilterBar(attrs),
         m(
           '.loading-container',
           m(
@@ -586,7 +256,6 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
         : undefined;
     return m(
       '.pf-flamegraph',
-      this.renderFilterBar(attrs),
       m(
         VirtualOverlayCanvas,
         {
@@ -736,49 +405,6 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
     );
   }
 
-  static createDefaultState(
-    metrics: ReadonlyArray<FlamegraphMetric>,
-  ): FlamegraphState {
-    return {
-      selectedMetricId: metricId(metrics[0]),
-      addedMetricIds: [],
-      filters: [],
-      view: {kind: 'TOP_DOWN'},
-    };
-  }
-
-  /**
-   * Updates a FlamegraphState with new metrics, preserving filters where possible.
-   *
-   * If the current state has no metric selected (empty string), this will
-   * initialize it with the first metric. Otherwise, it preserves the selected
-   * metric if it still exists in the new metrics array, or falls back to the
-   * first metric if it doesn't.
-   *
-   * Returns `state` unchanged (same reference) when it is already valid:
-   * callers rely on reference stability to avoid spurious data refetches.
-   */
-  static updateState(
-    state: FlamegraphState | undefined,
-    metrics: ReadonlyArray<FlamegraphMetric>,
-  ): FlamegraphState {
-    if (state === undefined) {
-      return Flamegraph.createDefaultState(metrics);
-    }
-    const metricStillExists = metrics.some(
-      (m) => metricId(m) === state.selectedMetricId,
-    );
-    if (metricStillExists) {
-      return state;
-    }
-    return {
-      filters: state.filters,
-      view: state.view,
-      addedMetricIds: state.addedMetricIds,
-      selectedMetricId: metricId(metrics[0]),
-    };
-  }
-
   private drawCanvas(
     ctx: CanvasRenderingContext2D,
     size: Size2D,
@@ -864,7 +490,8 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
         colorScheme = getFlamegraphColorScheme(name, state === 'PARTIAL');
       }
       const highlighted =
-        source.kind === 'NODE' && this.highlightRegex?.test(name) === true;
+        source.kind === 'NODE' &&
+        this.attrs.highlightRegex?.test(name) === true;
       const bgColor = hover ? colorScheme.variant : colorScheme.base;
       const textColor = hover ? colorScheme.textVariant : colorScheme.textBase;
       ctx.fillStyle = bgColor.cssString;
@@ -927,278 +554,6 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
     }
     const {y} = this.tooltipPos;
     return y >= this.viewportRect.top && y <= this.viewportRect.bottom;
-  }
-
-  private renderMeasurePicker(attrs: FlamegraphAttrs) {
-    const selected = attrs.metrics.find(
-      (metric) => metricId(metric) === attrs.state.selectedMetricId,
-    );
-    if (selected === undefined) {
-      return undefined;
-    }
-    const defaultMetrics = attrs.metrics.filter(
-      (metric) => metric.provenance === 'DEFAULT',
-    );
-    const otherMetrics = attrs.metrics.filter(
-      (metric) => metric.provenance !== 'DEFAULT',
-    );
-    const renderMetric = (metric: FlamegraphMetric) =>
-      m(MenuItem, {
-        label: metric.name,
-        rightIcon:
-          metricId(metric) === attrs.state.selectedMetricId
-            ? Icons.Check
-            : undefined,
-        onclick: () => {
-          attrs.onStateChange({
-            ...attrs.state,
-            selectedMetricId: metricId(metric),
-          });
-        },
-      });
-
-    return m(
-      PopupMenu,
-      {
-        trigger: m(Button, {
-          label: selected.name,
-          rightIcon: Icons.ExpandDown,
-        }),
-      },
-      defaultMetrics.map(renderMetric),
-      defaultMetrics.length > 0 && otherMetrics.length > 0 && m(MenuDivider),
-      otherMetrics.map(renderMetric),
-      attrs.addableMetrics !== undefined &&
-        attrs.addableMetrics.length > 0 && [
-          m(MenuDivider),
-          m(
-            MenuItem,
-            {label: 'Add measure...', icon: Icons.Add},
-            m(AddMetricMenu, {
-              metrics: attrs.addableMetrics,
-              onSelect: (metric) => attrs.onAddMetric?.(metric),
-            }),
-          ),
-        ],
-    );
-  }
-
-  private setHighlightPattern(pattern: string) {
-    this.highlightPattern = pattern;
-    if (pattern === '') {
-      this.highlightRegex = undefined;
-      return;
-    }
-    try {
-      const regex = parseUserFilterRegex(pattern);
-      this.highlightRegex = new RegExp(regex.pattern, regex.flags);
-    } catch {
-      this.highlightRegex = undefined;
-    }
-  }
-
-  private renderHighlightSearch(attrs: FlamegraphAttrs) {
-    const matchCount =
-      this.highlightRegex === undefined
-        ? 0
-        : (attrs.data?.nodes.filter((node) =>
-            this.highlightRegex?.test(node.name),
-          ).length ?? 0);
-    return m(
-      '.pf-flamegraph-highlight-search',
-      m(TextInput, {
-        autofocus: true,
-        leftIcon: Icons.Search,
-        placeholder: 'Name, /Regex/, or /regex/i…',
-        value: this.highlightPattern,
-        onInput: (value) => this.setHighlightPattern(value),
-        onkeydown: (event: KeyboardEvent) => {
-          if (event.key === 'Escape') {
-            this.setHighlightPattern('');
-          }
-        },
-      }),
-      this.highlightPattern !== '' &&
-        m(
-          'span.pf-flamegraph-highlight-search__count',
-          this.highlightRegex === undefined
-            ? 'Invalid regex'
-            : `${matchCount} ${matchCount === 1 ? 'match' : 'matches'}`,
-        ),
-    );
-  }
-
-  private renderFilterBar(attrs: FlamegraphAttrs) {
-    const tags = toTags(this.attrs.state);
-    const hasFilters = tags.length > 0;
-    const activePatternView =
-      attrs.state.view.kind === 'FROM_FRAME' ||
-      attrs.state.view.kind === 'PIVOT'
-        ? attrs.state.view.kind
-        : undefined;
-
-    const removeTag = (i: number) => {
-      if (i === this.attrs.state.filters.length) {
-        this.attrs.onStateChange({
-          ...this.attrs.state,
-          view: {kind: 'TOP_DOWN'},
-        });
-      } else {
-        const filters = this.attrs.state.filters.filter((_, j) => j !== i);
-        this.attrs.onStateChange({...this.attrs.state, filters});
-      }
-    };
-
-    const addFilterFn = (filters: Array<{type: FilterType; value: string}>) => {
-      let newState = this.attrs.state;
-      for (const {type, value} of filters) {
-        if (type === 'SHOW_FROM_FRAME') {
-          newState = {...newState, view: {kind: 'FROM_FRAME', pattern: value}};
-        } else if (type === 'PIVOT') {
-          newState = {...newState, view: {kind: 'PIVOT', pivot: value}};
-        } else {
-          newState = addFilter(newState, {kind: type, filter: value});
-        }
-      }
-      this.attrs.onStateChange(newState);
-    };
-
-    return m(
-      '.filter-bar',
-      m('span.pf-flamegraph-control-label', 'Measure:'),
-      this.renderMeasurePicker(attrs),
-      m('span.pf-flamegraph-control-label', 'Filters:'),
-      m(TagInput, {
-        tags,
-        value: this.quickAddValue,
-        onChange: (text) => {
-          this.quickAddValue = text;
-        },
-        onTagAdd: (text) => {
-          const filters = splitFilters(text).map((part) => parseFilter(part));
-          if (filters.length > 0) {
-            addFilterFn(filters);
-            this.quickAddValue = '';
-          }
-        },
-        onTagRemove: removeTag,
-        placeholder: hasFilters
-          ? ''
-          : 'e.g. malloc (contains), or /^main$/ for regex; press + for more filter options',
-        renderTag: (text, onRemove) =>
-          m(Chip, {
-            ondblclick: () => {
-              this.quickAddValue = text;
-              onRemove();
-            },
-            label: m(MiddleEllipsis, {text}),
-            removable: true,
-            compact: true,
-            intent: Intent.Primary,
-            onRemove,
-          }),
-      }),
-      m(
-        Popup,
-        {
-          trigger: m(Button, {
-            icon: Icons.Add,
-            compact: true,
-            active: this.showFilterBuilder,
-            onclick: () => {
-              this.showFilterBuilder = !this.showFilterBuilder;
-            },
-          }),
-          isOpen: this.showFilterBuilder,
-          onChange: (shouldOpen: boolean) => {
-            this.showFilterBuilder = shouldOpen;
-          },
-          position: PopupPosition.RightStart,
-          closeOnOutsideClick: true,
-          closeOnEscape: true,
-          className: 'pf-filter-builder',
-        },
-        m(FilterBuilder, {
-          activePatternView,
-          onAdd: addFilterFn,
-        }),
-      ),
-      m(CopyToClipboardButton(), {
-        textToCopy: () => tags.join(' '),
-        compact: true,
-        disabled: !hasFilters,
-      }),
-      m(Button, {
-        icon: 'delete',
-        compact: true,
-        disabled: !hasFilters,
-        onclick: () => {
-          attrs.onStateChange({
-            ...this.attrs.state,
-            filters: [],
-            view:
-              activePatternView === undefined
-                ? this.attrs.state.view
-                : {kind: 'TOP_DOWN'},
-          });
-        },
-      }),
-      m(
-        RadioGroup,
-        {
-          selectedValue:
-            attrs.state.view.kind === 'TOP_DOWN'
-              ? 'top-down'
-              : attrs.state.view.kind === 'BOTTOM_UP'
-                ? 'bottom-up'
-                : undefined,
-          onValueChange: (value) => {
-            attrs.onStateChange({
-              ...attrs.state,
-              view: {kind: value === 'top-down' ? 'TOP_DOWN' : 'BOTTOM_UP'},
-            });
-          },
-        },
-        [
-          m(RadioGroup.Button, {value: 'top-down'}, 'Top Down'),
-          m(RadioGroup.Button, {value: 'bottom-up'}, 'Bottom Up'),
-        ],
-      ),
-      m(Button, {
-        icon: Icons.Search,
-        label: 'Highlight',
-        active: this.showHighlightSearch || this.highlightPattern !== '',
-        onclick: () => {
-          this.showHighlightSearch = !this.showHighlightSearch;
-        },
-      }),
-      attrs.data !== undefined &&
-        attrs.data.nodes.length > 0 &&
-        m(ExportButton, {
-          fileBaseName: 'flamegraph',
-          onExportData: async (format) => this.buildExportString(format),
-        }),
-      this.showHighlightSearch &&
-        m(
-          '.pf-flamegraph-secondary-row',
-          m('span.pf-flamegraph-control-label', 'Highlight:'),
-          this.renderHighlightSearch(attrs),
-          this.highlightPattern !== '' &&
-            m(Button, {
-              icon: Icons.Close,
-              compact: true,
-              title: 'Clear highlight',
-              onclick: () => this.setHighlightPattern(''),
-            }),
-          m(Button, {
-            label: 'Hide',
-            compact: true,
-            onclick: () => {
-              this.showHighlightSearch = false;
-            },
-          }),
-        ),
-    );
   }
 
   private renderTooltip() {
@@ -1315,13 +670,13 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   private renderNodeActionBar(
     source: NodeSource,
     name: string,
-    node: FlamegraphNode,
-    nodeActions: ReadonlyArray<FlamegraphOptionalAction>,
+    node: TreeExplorerNode,
+    nodeActions: ReadonlyArray<TreeExplorerOptionalAction>,
   ) {
     const {properties} = node;
     const builtIn = this.buildNodeActions(source, name, node);
 
-    const isFlat = (a: FlamegraphOptionalAction) =>
+    const isFlat = (a: TreeExplorerOptionalAction) =>
       a.execute !== undefined &&
       (a.subActions === undefined || a.subActions.length === 0);
     const embedderFlat: NodeAction[] = nodeActions.filter(isFlat).map((a) => ({
@@ -1393,13 +748,13 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   private buildNodeActions(
     source: NodeSource,
     name: string,
-    node: FlamegraphNode,
+    node: TreeExplorerNode,
   ): NodeAction[] {
-    const applyState = (state: FlamegraphState) => {
+    const applyState = (state: TreeExplorerState) => {
       this.attrs.onStateChange(state);
       this.tooltipPos = undefined;
     };
-    const addF = (kind: FlamegraphFilter['kind'], filter: string) =>
+    const addF = (kind: TreeExplorerFilter['kind'], filter: string) =>
       applyState(addFilter(this.attrs.state, {kind, filter}));
     // Match this exact name: an anchored regex over the escaped literal name,
     // wrapped in `/…/` so the filter bar reads it as a regex.
@@ -1497,8 +852,8 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
 
   // Root-only actions menu; node actions go through renderNodeActionBar.
   private renderActionsMenu(
-    actions: ReadonlyArray<FlamegraphOptionalAction>,
-    properties: ReadonlyMap<string, FlamegraphPropertyDefinition>,
+    actions: ReadonlyArray<TreeExplorerOptionalAction>,
+    properties: ReadonlyMap<string, TreeExplorerPropertyDefinition>,
   ) {
     if (actions.length === 0) {
       return null;
@@ -1518,9 +873,9 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   }
 
   private renderMenuItem(
-    action: FlamegraphOptionalAction,
-    properties: ReadonlyMap<string, FlamegraphPropertyDefinition>,
-    node?: FlamegraphNode,
+    action: TreeExplorerOptionalAction,
+    properties: ReadonlyMap<string, TreeExplorerPropertyDefinition>,
+    node?: TreeExplorerNode,
   ): m.Vnode<MenuItemAttrs> {
     if (action.subActions !== undefined && action.subActions.length > 0) {
       return this.renderParentMenuItem(
@@ -1537,10 +892,10 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   }
 
   private renderParentMenuItem(
-    action: FlamegraphOptionalAction,
-    subActions: FlamegraphOptionalAction[],
-    properties: ReadonlyMap<string, FlamegraphPropertyDefinition>,
-    node?: FlamegraphNode,
+    action: TreeExplorerOptionalAction,
+    subActions: TreeExplorerOptionalAction[],
+    properties: ReadonlyMap<string, TreeExplorerPropertyDefinition>,
+    node?: TreeExplorerNode,
   ): m.Vnode<MenuItemAttrs> {
     return m(
       MenuItem,
@@ -1557,9 +912,9 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   }
 
   private renderExecutableMenuItem(
-    action: FlamegraphOptionalAction,
-    properties: ReadonlyMap<string, FlamegraphPropertyDefinition>,
-    node?: FlamegraphNode,
+    action: TreeExplorerOptionalAction,
+    properties: ReadonlyMap<string, TreeExplorerPropertyDefinition>,
+    node?: TreeExplorerNode,
   ): m.Vnode<MenuItemAttrs> {
     return m(MenuItem, {
       label: this.actionItemLabel(action.name, action.description),
@@ -1575,7 +930,7 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
   }
 
   private renderDisabledMenuItem(
-    action: FlamegraphOptionalAction,
+    action: TreeExplorerOptionalAction,
   ): m.Vnode<MenuItemAttrs> {
     return m(MenuItem, {
       label: this.actionItemLabel(action.name, action.description),
@@ -1584,13 +939,16 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
     });
   }
 
-  private buildStackString(node: FlamegraphNode, withDetails: boolean): string {
+  private buildStackString(
+    node: TreeExplorerNode,
+    withDetails: boolean,
+  ): string {
     const {nodes, unfilteredCumulativeValue} = ensureExists(this.attrs.data);
     const metric = ensureExists(this.selectedMetric);
     const view = this.attrs.state.view;
 
     // Walk via parentId for all modes. Reverse for TOP_DOWN and PIVOT below.
-    const stack: FlamegraphNode[] = [];
+    const stack: TreeExplorerNode[] = [];
     let currentId = node.id;
     while (currentId !== -1) {
       const current = ensureExists(nodes.find((n) => n.id === currentId));
@@ -1686,83 +1044,86 @@ export class Flamegraph implements m.ClassComponent<FlamegraphAttrs> {
     return lines.join('\n');
   }
 
-  // Builds a table of all currently displayed nodes (i.e. with filters
-  // applied) in the given format. Metric values are raw numbers in the
-  // metric's unit so they can be aggregated in spreadsheets; the tree
-  // structure is preserved via the id/parentId columns.
-  private buildExportString(format: ExportFormat): string {
-    const {nodes} = ensureExists(this.attrs.data);
-    const metric = ensureExists(this.selectedMetric);
-    const unitDisplay = getUnitDisplayName(metric.unit);
-
-    const unaggKeys: string[] = [];
-    const aggKeys: string[] = [];
-    const propDisplayNames = new Map<string, string>();
-    for (const node of nodes) {
-      for (const [key, prop] of node.properties) {
-        const keys = prop.isAggregatable ? aggKeys : unaggKeys;
-        if (!keys.includes(key)) {
-          keys.push(key);
-          propDisplayNames.set(key, prop.displayName);
-        }
-      }
-    }
-
-    const columns = [
-      'id',
-      'parentId',
-      'depth',
-      'name',
-      ...unaggKeys,
-      'cumulativeValue',
-      'selfValue',
-      ...aggKeys,
-    ];
-    const columnNames: Record<string, string> = {
-      ...Object.fromEntries(propDisplayNames),
-      id: 'Id',
-      parentId: 'Parent Id',
-      depth: 'Depth',
-      name: metric.nameColumnLabel ?? 'Name',
-      cumulativeValue: `Cumulative ${metric.name} (${unitDisplay})`,
-      selfValue: `Self ${metric.name} (${unitDisplay})`,
-    };
-    const rows = nodes.map((n) => {
-      const row: Record<string, string> = {
-        id: n.id.toString(),
-        parentId: n.parentId.toString(),
-        depth: n.depth.toString(),
-        name: n.name,
-        cumulativeValue: n.cumulativeValue.toString(),
-        selfValue: n.selfValue.toString(),
-      };
-      for (const key of [...unaggKeys, ...aggKeys]) {
-        row[key] = n.properties.get(key)?.value ?? '';
-      }
-      return row;
-    });
-
-    switch (format) {
-      case 'tsv':
-        return formatAsTSV(columns, columnNames, rows);
-      case 'json':
-        return formatAsJSON(columns, columnNames, rows);
-      case 'markdown':
-        return formatAsMarkdown(columns, columnNames, rows);
-      default:
-        assertUnreachable(format);
-    }
-  }
-
   private createReducedProperties(
-    properties: ReadonlyMap<string, FlamegraphPropertyDefinition>,
+    properties: ReadonlyMap<string, TreeExplorerPropertyDefinition>,
   ): ReadonlyMap<string, string> {
     return new Map([...properties].map(([key, {value}]) => [key, value]));
   }
 }
 
+// Builds a table of all currently displayed nodes (i.e. with filters
+// applied) in the given format. Metric values are raw numbers in the
+// metric's unit so they can be aggregated in spreadsheets; the tree
+// structure is preserved via the id/parentId columns.
+export function buildFlamegraphExportString(
+  data: TreeExplorerData,
+  metric: TreeExplorerMetric,
+  format: ExportFormat,
+): string {
+  const {nodes} = data;
+  const unitDisplay = getUnitDisplayName(metric.unit);
+
+  const unaggKeys: string[] = [];
+  const aggKeys: string[] = [];
+  const propDisplayNames = new Map<string, string>();
+  for (const node of nodes) {
+    for (const [key, prop] of node.properties) {
+      const keys = prop.isAggregatable ? aggKeys : unaggKeys;
+      if (!keys.includes(key)) {
+        keys.push(key);
+        propDisplayNames.set(key, prop.displayName);
+      }
+    }
+  }
+
+  const columns = [
+    'id',
+    'parentId',
+    'depth',
+    'name',
+    ...unaggKeys,
+    'cumulativeValue',
+    'selfValue',
+    ...aggKeys,
+  ];
+  const columnNames: Record<string, string> = {
+    ...Object.fromEntries(propDisplayNames),
+    id: 'Id',
+    parentId: 'Parent Id',
+    depth: 'Depth',
+    name: metric.nameColumnLabel ?? 'Name',
+    cumulativeValue: `Cumulative ${metric.name} (${unitDisplay})`,
+    selfValue: `Self ${metric.name} (${unitDisplay})`,
+  };
+  const rows = nodes.map((n) => {
+    const row: Record<string, string> = {
+      id: n.id.toString(),
+      parentId: n.parentId.toString(),
+      depth: n.depth.toString(),
+      name: n.name,
+      cumulativeValue: n.cumulativeValue.toString(),
+      selfValue: n.selfValue.toString(),
+    };
+    for (const key of [...unaggKeys, ...aggKeys]) {
+      row[key] = n.properties.get(key)?.value ?? '';
+    }
+    return row;
+  });
+
+  switch (format) {
+    case 'tsv':
+      return formatAsTSV(columns, columnNames, rows);
+    case 'json':
+      return formatAsJSON(columns, columnNames, rows);
+    case 'markdown':
+      return formatAsMarkdown(columns, columnNames, rows);
+    default:
+      assertUnreachable(format);
+  }
+}
+
 function computeRenderNodes(
-  {nodes, allRootsCumulativeValue, minDepth}: FlamegraphQueryData,
+  {nodes, allRootsCumulativeValue, minDepth}: TreeExplorerData,
   zoomRegion: ZoomRegion,
   canvasWidth: number,
 ): ReadonlyArray<RenderNode> {
@@ -1923,232 +1284,6 @@ function isIntersecting(
     needleY >= y &&
     needleY < y + NODE_HEIGHT
   );
-}
-
-function metricId(metric: FlamegraphMetric): string {
-  return metric.id ?? metric.name;
-}
-
-interface AddMetricMenuAttrs {
-  readonly metrics: ReadonlyArray<FlamegraphAddableMetric>;
-  readonly onSelect: (metric: FlamegraphAddableMetric) => void;
-}
-
-class AddMetricMenu implements m.ClassComponent<AddMetricMenuAttrs> {
-  private static readonly MAX_VISIBLE_ITEMS = 100;
-  private searchQuery = '';
-
-  view({attrs}: m.CVnode<AddMetricMenuAttrs>) {
-    const results =
-      this.searchQuery === ''
-        ? attrs.metrics.map((metric) => ({
-            metric,
-            segments: [{matching: false, value: metric.name}],
-          }))
-        : fuzzySearch(
-            attrs.metrics,
-            (metric) => metric.name,
-            this.searchQuery,
-          ).map((result) => ({
-            metric: result.item,
-            segments: result.segments,
-          }));
-    const visible = results.slice(0, AddMetricMenu.MAX_VISIBLE_ITEMS);
-    const remaining = results.length - visible.length;
-
-    return m('.pf-distinct-values-menu', [
-      m(
-        '.pf-distinct-values-menu__search',
-        {
-          onclick: (event: MouseEvent) => event.stopPropagation(),
-        },
-        m(TextInput, {
-          placeholder: 'Search measures...',
-          value: this.searchQuery,
-          oninput: (event: InputEvent) => {
-            this.searchQuery = (event.target as HTMLInputElement).value;
-          },
-          onkeydown: (event: KeyboardEvent) => {
-            if (this.searchQuery !== '' && event.key === 'Escape') {
-              this.searchQuery = '';
-              event.stopPropagation();
-            }
-          },
-        }),
-      ),
-      m(
-        '.pf-distinct-values-menu__list',
-        visible.length > 0
-          ? [
-              visible.map(({metric, segments}) =>
-                m(MenuItem, {
-                  label: segments.map((segment) =>
-                    segment.matching
-                      ? m('strong.pf-fuzzy-match', segment.value)
-                      : segment.value,
-                  ),
-                  onclick: () => {
-                    attrs.onSelect(metric);
-                    this.searchQuery = '';
-                  },
-                }),
-              ),
-              remaining > 0 &&
-                m(MenuItem, {
-                  label: `...and ${remaining} more`,
-                  disabled: true,
-                }),
-            ]
-          : m(EmptyState, {title: 'No matches'}),
-      ),
-    ]);
-  }
-}
-
-function displaySize(totalSize: number, unit: string): string {
-  if (unit === '' || unit === 'count') return totalSize.toLocaleString();
-  if (totalSize === 0) return `0 ${unit}`;
-  let step: number;
-  let units: string[];
-  switch (unit) {
-    case 'B':
-      step = 1024;
-      units = ['B', 'KiB', 'MiB', 'GiB'];
-      break;
-    case 'ns':
-      step = 1000;
-      units = ['ns', 'us', 'ms', 's'];
-      break;
-    default:
-      step = 1000;
-      units = [unit, `K${unit}`, `M${unit}`, `G${unit}`];
-      break;
-  }
-  const unitsIndex = Math.min(
-    Math.trunc(Math.log(totalSize) / Math.log(step)),
-    units.length - 1,
-  );
-  const pow = Math.pow(step, unitsIndex);
-  const result = totalSize / pow;
-  const resultString =
-    totalSize % pow === 0 ? result.toString() : result.toFixed(2);
-  return `${resultString} ${units[unitsIndex]}`;
-}
-
-function displayPercentage(size: number, totalSize: number): string {
-  if (totalSize === 0) {
-    return `[NULL]%`;
-  }
-  return `${((size / totalSize) * 100.0).toFixed(2)}%`;
-}
-
-function getUnitDisplayName(unit: string | undefined): string {
-  if (unit === undefined || unit === '' || unit === 'count') {
-    return 'count';
-  }
-  return unit;
-}
-
-function toTags(state: FlamegraphState): ReadonlyArray<string> {
-  const toString = (x: FlamegraphFilter) => {
-    switch (x.kind) {
-      case 'HIDE_FRAME':
-        return 'Hide Frame: ' + x.filter;
-      case 'HIDE_STACK':
-        return 'Hide Stack: ' + x.filter;
-      case 'SHOW_STACK':
-        return 'Show Stack: ' + x.filter;
-      case 'OPTIONS':
-        return 'Options';
-    }
-  };
-  const filters = state.filters.map((x) => toString(x));
-  switch (state.view.kind) {
-    case 'FROM_FRAME':
-      return filters.concat([
-        'Show From Frame: ' + (state.view.displayLabel ?? state.view.pattern),
-      ]);
-    case 'PIVOT':
-      return filters.concat([
-        'Pivot: ' + (state.view.displayLabel ?? state.view.pivot),
-      ]);
-    case 'TOP_DOWN':
-    case 'BOTTOM_UP':
-      return filters;
-    default:
-      assertUnreachable(state.view);
-  }
-}
-
-function addFilter(
-  state: FlamegraphState,
-  filter: FlamegraphFilter,
-): FlamegraphState {
-  return {
-    ...state,
-    filters: state.filters.concat([filter]),
-  };
-}
-
-// Split text into individual filters by finding filter type prefixes
-// e.g. 'Show Stack: main Hide Frame: alloc' -> ['Show Stack: main', 'Hide Frame: alloc']
-// e.g. 'SS: foo HF: bar' -> ['SS: foo', 'HF: bar']
-function splitFilters(text: string): string[] {
-  const lowerText = text.toLowerCase();
-
-  // Find all positions where a filter prefix starts (case insensitive)
-  const splitPositions: number[] = [];
-  for (const type of FILTER_TYPES) {
-    for (const prefix of [type.shortLabel, type.label]) {
-      const searchStr = prefix.toLowerCase() + ':';
-      let pos = 0;
-      while ((pos = lowerText.indexOf(searchStr, pos)) !== -1) {
-        // Only split if at start or preceded by whitespace
-        if (pos === 0 || /\s/.test(text[pos - 1])) {
-          splitPositions.push(pos);
-        }
-        pos += searchStr.length;
-      }
-    }
-  }
-
-  // Sort and deduplicate positions
-  splitPositions.sort((a, b) => a - b);
-
-  // If no prefixes found, return the whole text as one filter
-  if (splitPositions.length === 0) {
-    return text.trim() ? [text.trim()] : [];
-  }
-
-  // Split text at those positions
-  const result: string[] = [];
-  for (let i = 0; i < splitPositions.length; i++) {
-    const start = splitPositions[i];
-    const end = splitPositions[i + 1] ?? text.length;
-    const part = text.substring(start, end).trim();
-    if (part) {
-      result.push(part);
-    }
-  }
-  return result;
-}
-
-// Parse a filter string into type and value
-// e.g. 'SS: main' -> {type: 'SHOW_STACK', value: 'main'}
-// e.g. 'Show Stack: main' -> {type: 'SHOW_STACK', value: 'main'}
-function parseFilter(
-  text: string,
-  defaultType: FilterType = 'SHOW_STACK',
-): {type: FilterType; value: string} {
-  const i = text.indexOf(':');
-  if (i === -1) return {type: defaultType, value: text};
-  const prefix = text.substring(0, i).trim().toLowerCase();
-  const value = text.substring(i + 1).trim();
-  const match = FILTER_TYPES.find(
-    (o) =>
-      o.shortLabel.toLowerCase() === prefix || o.label.toLowerCase() === prefix,
-  );
-  return match ? {type: match.value, value} : {type: defaultType, value: text};
 }
 
 // Unfortunately, widgets *cannot* depend on components so we cannot use the
